add iconv-lite to supported encodings
